* Fixed compilation error in t14 * Modified all CMakeLists.txt to produce nicer output
CCMAIL: kde-bindings@kde.org


svn path=/trunk/playground/bindings/kimono/; revision=633652

+++ qyoto/examples/tutorial/t14/gameboard.cs
@@ -45,7 +45,7 @@ class GameBoard : QWidget {
         force.SetRange(10, 50);
 
         QFrame cannonBox = new QFrame();
-        cannonBox.SetFrameStyle((int) (QFrame.Shape.WinPanel | QFrame.Shadow.Sunken));
+        cannonBox.SetFrameStyle((int) QFrame.Shape.WinPanel | (int) QFrame.Shadow.Sunken);
 
         cannonField = new CannonField();
